Beta 1.16.5-0.1.8 -Ported from 1.19.4-0.1.7 -Changed energy reborn API version to 0.1.1 -Removed for now Crafty Energy Capability in Forge -and a lot of minor changes for 1.16.5 version adaptation...

+++ common/src/main/java/wily/factoryapi/base/IPlatformFluidHandler.java
@@ -1,6 +1,6 @@
 package wily.factoryapi.base;
 
-import dev.architectury.fluid.FluidStack;
+import me.shedaniel.architectury.fluid.FluidStack;
 import net.minecraft.nbt.CompoundTag;
 import org.jetbrains.annotations.NotNull;
 
@@ -79,7 +79,7 @@ public interface IPlatformFluidHandler<T> extends ITagSerializable<CompoundTag>,
     FluidStack drain(int maxDrain, boolean simulate);
 
     default long getTotalSpace(){
-        return Math.max(0, getMaxFluid() - getFluidStack().getAmount());
+        return Math.max(0, getMaxFluid() - getFluidStack().getAmount().longValue());
     }
